State event for steady state detection was corrected (returning float instead of int for the combined steady state criteria). Also, correct units for the moments for plots were included.

+++ PharmaPy/Commons.py
@@ -239,7 +239,7 @@ def check_steady_state(time, states, sdot, tau, num_tau=1, time_stop=None,
 
     flag = not all(flags)
 
-    return int(flag)
+    return float(flag)
 
 
 def eval_state_events(time, states, switches,
